Add files via upload
# Iris Flower Prediction Project

Welcome to my Iris Flower Prediction project repository! 🌼 In this project, I've built a machine learning model that predicts the species of Iris flowers based on their characteristics. The Iris dataset is a classic in the world of data science and machine learning, making it an excellent learning experience for both beginners and enthusiasts.

## Key Features:

- Jupyter Notebook: Dive into the code and explore the step-by-step process of building the model.
- Machine Learning: Get insights into the machine learning techniques used for classification.
- Data Exploration: Discover the exploratory data analysis performed on the Iris dataset.
- Model Evaluation: Understand the model's performance and how it predicts Iris species.

This project is a great starting point for anyone interested in machine learning and data science.  Enjoy your journey with the Iris Flower Prediction project! 🌱